requirements.txt
attrs==23.2.0
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
confluent-kafka==2.4.0
exceptiongroup==1.2.1
faker==26.0.0
flask==3.0.3
flask-sqlalchemy==3.1.1
greenlet==3.0.3
idna==3.7
importlib-metadata==7.2.1
iniconfig==2.0.0
itsdangerous==2.2.0
jinja2==3.1.4
jsonschema==3.2.0
markupsafe==2.1.5
mysqlclient==2.0.3
packaging==24.1
pluggy==1.5.0
pyrsistent==0.20.0
pytest==8.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sqlalchemy==2.0.31
tomli==2.0.1
typing-extensions==4.12.2
urllib3==2.2.2
werkzeug==3.0.3
zipp==3.19.2